Add note for possible GCloud CLI setup issue
This has happened twice to me now where I have set up the dev environment as described, but I did not update my GCloud CLI component. Running `npm start` suppresses the error message that states to update the component before continuing, and an unrelated error will appear. This PR adds this corner case to the README notes just in case it is a problem for anyone else.